Add .mailmap for commit author translations
Because it's not possible to enumerate up-front all of the committers to
tmux coming from OpenBSD, at the time a commit is imported in to git from
the OpenBSD CVS repository, the author information is not known to Git,
necessarily.

But it's possible to alter for output the respective author after the fact,
via Git's .mailmap file.  It is this file which will therefore provide a new
mapping of OpenBSD commiter to an actual real name and real email address.