Add new terminal-aliases server option
The option adds a mapping between terminal settings in $TERM variable
and the actual terminfo entry to use.

It can be more effective than terminal-overrides for tasks such as
enabling 256 colour support, since that usually requires setting more
than a single terminfo variable to the correct values.

For instance, until tmux 1.8 this recipe used to work to enable
256 colour support on xterm:

  set -g terminal-overrides "xterm:colors=256"

However, commit f2e54e1e2ff5a1 started using "setaf" and "setab"
terminfo settings, which are insufficient in "xterm" to set 256 colours.
It is possible to override those, but if the goal is to enable support
for 256 colours, it is more effective to do so by instructing tmux to
just use the full "xterm-256color" terminfo entry directly, which can
now be achieved using:

  set -g terminal-aliases "xterm:xterm-256color"

Of course, ideally the $TERM setting should be correct to specify
"xterm-256color" if that support is available, but unfortunately some
terminal emulators are limited (for instance, xfce4-terminal uses
termcap which has only a basic set of entries these days) and there are
number of systems misconfigured to set TERM=xterm or similar from global
configuration files such as /etc/profile, which often makes it more
convenient to just override settings in tmux configuration (same as was
already done, to quite an extent, by terminal-overrides, anyways.)

Tested on xfce4-terminal, using the "xterm:xterm-256color" mapping,
confirming that vim syntax highlighting with 256 colour support was
available inside the freshly started tmux session. Man page updated
and tested locally.

Signed-off-by: Filipe Brandenburger <filbranden@gmail.com>