Add support for a use=... in terminal-overrides to perform terminal name mapping.
This allows creating a mapping between the name in $TERM variable and
the actual terminfo entry to use.

Mapping a full terminfo entry can be more effective than overriding
individual variables in terminal-overrides, particularly for features
such as getting 256 colour support to work, since it often requires
overriding multiple terminfo variables.

For instance, until tmux 1.8 this recipe used to work to enable
256 colour support on xterm:

  set -g terminal-overrides "xterm:colors=256"

However, commit f2e54e1e2ff5a1 started using the "setaf" and "setab"
terminfo variables to set colors, and the values of those variables for
the "xterm" terminal are insufficient for 256 colour support.

It is possible to override those individually, but if the goal is to
enable 256 colours, it is more effective to do so by simply telling tmux
to use "xterm-256color" directly, which can now be achieved using:

  set -g terminal-overrides "xterm:use=xterm-256color"

Ideally, this should be fixed by setting $TERM correct to directly
specify "xterm-256color" if 256 colours are indeed available.
Unfortunately, some terminal emulators are limited (for instance,
xfce4-terminal's VTE only supports a limited set of entries in which
"xterm-256color" is not included) and some systems are misconfigured to
set TERM=xterm unconditionally from global configuration files such as
/etc/profile, which often makes it more convenient to just override
settings in tmux configuration. In a way, terminal-overrides was already
serving this purpose, this only extends it to allow it to map a whole
terminal setting completely.

Tested on xfce4-terminal, using the "xterm:use=xterm-256color:XT"
mapping, confirmed that vim syntax highlighting with 256 colour support
was available inside the freshly started tmux session.

Signed-off-by: Filipe Brandenburger <filbranden@gmail.com>