Fix fill in code to store boolean flags in the .flag member.
In practice, the end result was correct, as this is the field of an
union and both .number and .flag point to the same memory. But certainly
using .flag for TTYCODE_FLAG is semantically correct and prevents
confusion, so we should fix that.

Tested by adding debug logs to confirm that boolean fields were still
properly populated after this fix.

Signed-off-by: Filipe Brandenburger <filbranden@gmail.com>