Cache commands executed with #() and default to 1 for status-interval
As of 379400cfa69f8df9ac13b070c60d5f8a282ddf6e, commands have been run
every second when the status bar is being redrawn (on every event loop).

This commit introduces a ‘command-cache-time’ setting that controls the
caching time of #() commands and reduces status-interval to 1, so clocks
are redrawn in time and not only when the event loop happens to be
executed.