Separates ncurses search from other libraries
On my system when PKG_CHECK_MODULES fails the fall back succeed to find setupterm into tinfo library printing no error even if ncurses.h is not installed. This patch adds a specific test for addch to be present into libncurses. It returns an error when ncurses.h is not installed.