Added -N option to new-session
-N does the following:

1. Creates a group with the name specified on the command
   line, if one doesn't exist yet
2. If there are any unattached sessions in the named group
   then those will be attached to, otherwise a new group is made.

The benefit is that you can run a command like:

tmux new-session -N test

Which will launch a new session if one doesnt exist, or will
attach to an existing session if one does, this replaces the
common 'tmux attach || tmux' alias.  It also extends it so that
the same command can launch any number of sessions in the same
group, allowing the easy use of multiple terminal emulators on
the same group.