Add full-size window split feature
This adds a new argument (-f) to the split-window command which causes
the new pane to be created spanning the full window height (with -h) or
full window width (with -v) instead of splitting the active pane.

When creating a new full-size split, all existing panes are resized
proportionally to fit within the reduced window space. The command fails
if there is not sufficient space available in the window.

To illustrate, split-window -h on the following layout (with pane 1
active) produces:

    +---------+      +----+----+
    | 1       |      | 1  | 3  |
    +---------+  ->  +----+----+
    | 2       |      | 2       |
    +---------+      +---------+

With this feature, split-window -h -f on the same layout produces:

    +---------+      +----+----+
    | 1       |      | 1  | 3  |
    +---------+  ->  +----+    |
    | 2       |      | 2  |    |
    +---------+      +---------+