Support multiplexing SSH-agent sessions
Add a feature to multiplex other systems through the same connectivity as tmux
windows and clients. This creates a SSH agent for each window and forwards
requests to the SSH agents discovered by clients, then synthesises the
responses. There is a unique agent for each window and the current
window-client connections determine which forwarding is valid.

This is meant to be extensible to other protocols, including GPG agent and
DBus.