Implement ccc on a per-pane basis
Each pane has a 256color palette.  Sending the OSC sequence

    \x1b]4;d;rgb:r/g/b\x1b\\

sets the dth palette entry to the RGB color denoted by r, g, b and redraws
the pane. In tty_colours_{bg,fg}, this palette is used for substitution
before the actual color is set, so that palette changing is implemented
via 24-bit color support.

The special-case logic in tty_colours is slightly altered to allow
fg/bg values of 8 to match the pane-specific palette instead of the host
terminal's defaults.