compat: don't re-define program_invocation_short_name
program_invocation_short_name is defined in errno.h, and its definition
can differ between the various C libraries: glibc defines it as:
    extern char *program_invocation_short_name;

while uClibc defines it as:
    extern const char *program_invocation_short_name;

So there is not simple solution to know the prototype.

But since it is defined in errno.h, there is no reason to try and define
it ourselves; let's just trust what the header provides.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>