Interpret `tmux source-file` path relative to $PWD
Previously, they were interpreted relative to the server's current
directory, which was surprising when using tmux from the command line:

    $ cd ~ && tmux  # following commands within tmux session
    $ cd ~/.ssh
    $ file ../.tmux.conf
    /home/$USER/.tmux.conf: ASCII text
    $ tmux source-file ../.tmux.conf
    ../.tmux.conf: No such file or directory

This change considers the current directory when calling tmux from the
command-line. As a bonus, it lets you mark scripts for tmux with a
shebang:

    #!/usr/bin/env tmux source-file

Previously, you could do that and the file would work if called with an
absolute path or from the tmux server's $PWD, but not in general.

Implementation notes:

  * The code to determine the appropriate $PWD is copied from
    cmd-load-buffer.c. I'm not sure what the difference between c->cwd
    and c->session->cwd is, but I assume that load-buffer is correct on
    it.

  * I've preserved the existing behavior of -q, which interprets
    globbing failure differently from sourcing failure. For example:

        $ tmux source-file -q /etc/sudoers && echo yes || echo no
        yes
        $ tmux source-file -q ~root/.tmux.conf && echo yes || echo no
        no

    This happens because sudoers can be globbed but not read, whereas
    glob can't determine if ~root/.tmux.conf exists or not.