handle pty devices from different namespaces
Various programs that deal with namespaces will use pty devices that exist in
another namespace. One obvious candiate are containers. So far ttyname() was
incorrectly handling this case because the pts device stems from the
host and thus cannot be found amongst the current namespace's /dev/pts/<n>
entries. Serge Hallyn and I recently upstreamed patches to glibc that allow
ttyname{_r}() to correctly handle this case. At a minimum, ttyname{_r}() will
set errno to ENODEV in case it finds that the /dev/pts/<n> device that the
symlink points to exists in another namespace. This commit will allow tmux to
handle this case and behave correctly in a container.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>