Add OSC 11 support for changing background color
Add support for OSC 11 escape sequences to allow the background color of
the current pane to be changed. For example, the following changes the
background color to red:
   printf "\033]11;rgb:ff/00/00\033\\"

This allows shell color scheme scripts such as base16-shell to set
themes with correct background colors for individual panes. Note, in
order to set the color scheme only for the current pane, base16-shell
scripts should be called with "env TMUX= <script path>" to avoid sending
the escape sequence which tells tmux to pass the escape sequence through
to the terminal.