Compact line's grid extended data when it enters history
When working with a 50,000 line history, heavy with 24-bit color ANSI,
I grew suspicious noticing the memory resident size of the tmux being
1.6GB. Some debugging discovered that the `extdsize` of many history
lines was much greater than their `cellsize`. Regardless of the cause
of that, this commit makes sure that the history's memory usage is
kept closer to the optimum. After this fix, the memory resident size
of tmux under the same 50,000 line history scenario dropped down
to 170MB.