Faster grid reflow for vertical resize with many split lines
For large 200,000 lines pane histories which contain split lines,
the latency in vertical pane resize is noticable due to joins
and re-splits in the grid reflow process.

If we take into account that the pane's horizontal size has not
changed, we can just move those lines from the old grid instead
of spending time joining and re-spliting them.