try to fix wrapping when having a two line zsh prompt
On SIGWINCH zsh sometimes rewrites its prompt 'further up' on the
screen. This may put the rewritten output directly below a line with the
'wrapped' flag set. On subsequent reflow the wrapping algorithm may try
to wrap the old, wrapped line with the rewritten output. Unset the
wrapped flag to prevent this.