Fix resizing of panes by mouse
The iteration of the list pane was not reliable in finding the right
pane to resize. Moreover, we actually want to find the inter-pane
border that is relevant to the resize action.

Therefore, instead of iterating the list of panes, go over the list
of layouts in order to find the window that is next to the border that
we are trying to resize. The behavior of a pane border drag is to
always move the border, regardless of pane sizes (as long as they don't
go over the minimum)

This addresses issue #1002.