mouse events: ignore pane borders while zoomed
Previously, when creating a mouse event and settings its window pane,
the pane-search procedure tested pane borders unconditionally, and if
the coordinates were on a border, then this border's pane was set for
the event.

The issue is that in zoom mode borders are irrelevant, and it was
possible to end up with a pane which isn't the active (zoomed) one.

This resulted in mouse events getting ignored/other-issues in zoom mode
if they happen where a pane border exists when unzoomed.

This commit simply skips the border search procedure if the window is
zoomed.

Fixes #1473